requirements.txt
pyobjc-core==2.5.1
pyobjc-framework-Cocoa==2.5.1
